+++ src/main/java/svnserver/ext/gitlfs/server/LfsContentManager.java
@@ -41,6 +41,7 @@ import java.util.Objects;
  */
 public class LfsContentManager implements ContentManager {
   private final int tokenExpireSec;
+  private final int tokenEnsureSec;
 
   @FunctionalInterface
   public interface Checker {
@@ -52,15 +53,16 @@ public class LfsContentManager implements ContentManager {
   @NotNull
   private final LfsStorage storage;
 
-  public LfsContentManager(@NotNull LocalContext context, @NotNull LfsStorage storage, int tokenExpireSec) {
+  public LfsContentManager(@NotNull LocalContext context, @NotNull LfsStorage storage, int tokenExpireSec, int tokenEnsureSec) {
     this.context = context;
     this.storage = storage;
     this.tokenExpireSec = tokenExpireSec;
+    this.tokenEnsureSec = tokenEnsureSec;
   }
 
   private User getAuthInfo(@NotNull HttpServletRequest request) {
     final WebServer server = context.getShared().sure(WebServer.class);
-    final User user = server.getAuthInfo(request.getHeader(Constants.HEADER_AUTHORIZATION));
+    final User user = server.getAuthInfo(request.getHeader(Constants.HEADER_AUTHORIZATION), tokenEnsureSec);
     return user == null ? User.getAnonymous() : user;
   }